+++ squirls-cli/src/main/java/org/monarchinitiative/squirls/cli/visualization/MissingFeatureException.java
@@ -1,4 +1,4 @@
-package org.monarchinitiative.squirls.cli.cmd.analyze_vcf.visualization;
+package org.monarchinitiative.squirls.cli.visualization;
 
 import org.monarchinitiative.squirls.core.SquirlsException;