+++ squirls-cli/src/test/java/org/monarchinitiative/squirls/cli/visualization/simple/VisualizationContextSelectorTest.java
@@ -1,4 +1,4 @@
-package org.monarchinitiative.squirls.cli.cmd.analyze_vcf.visualization.simple;
+package org.monarchinitiative.squirls.cli.visualization.simple;
 
 import de.charite.compbio.jannovar.annotation.VariantAnnotator;
 import de.charite.compbio.jannovar.data.JannovarData;
@@ -6,8 +6,11 @@ import de.charite.compbio.jannovar.data.ReferenceDictionary;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.monarchinitiative.squirls.cli.TestDataSourceConfig;
-import org.monarchinitiative.squirls.cli.cmd.analyze_vcf.SplicingVariantAlleleEvaluation;
+import org.monarchinitiative.squirls.cli.cmd.analyze_vcf.data.SplicingVariantAlleleEvaluation;
 import org.monarchinitiative.squirls.cli.data.VariantsForTesting;
+import org.monarchinitiative.squirls.cli.visualization.SimpleVisualizationContextSelector;
+import org.monarchinitiative.squirls.cli.visualization.VisualizationContext;
+import org.monarchinitiative.squirls.cli.visualization.VisualizationContextSelector;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -30,7 +33,7 @@ public class VisualizationContextSelectorTest {
     @BeforeEach
     public void setUp() {
         rd = jannovarData.getRefDict();
-        selector = new VisualizationContextSelector();
+        selector = new SimpleVisualizationContextSelector();
     }
 
     // ****************************************** DONOR ****************************************************************
@@ -38,21 +41,21 @@ public class VisualizationContextSelectorTest {
     @Test
     public void donorQuidVariant() throws Exception {
         final SplicingVariantAlleleEvaluation ve = VariantsForTesting.BRCA2DonorExon15plus2QUID(rd, annotator);
-        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction());
+        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction().getFeatureMap());
         assertThat(ctx, is(VisualizationContext.CANONICAL_DONOR));
     }
 
     @Test
     public void donorNonQuidVariantThatDisruptsTheSite() throws Exception {
         final SplicingVariantAlleleEvaluation ve = VariantsForTesting.ALPLDonorExon7Minus2(rd, annotator);
-        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction());
+        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction().getFeatureMap());
         assertThat(ctx, is(VisualizationContext.CANONICAL_DONOR));
     }
 
     @Test
     public void donorNonQuidVariantThatCreatesACrypticDonor() throws Exception {
         final SplicingVariantAlleleEvaluation ve = VariantsForTesting.HBBcodingExon1UpstreamCrypticInCanonical(rd, annotator);
-        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction());
+        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction().getFeatureMap());
         assertThat(ctx, is(VisualizationContext.CRYPTIC_DONOR));
     }
 
@@ -60,7 +63,7 @@ public class VisualizationContextSelectorTest {
     @Test
     public void codingNonQuidVariantThatCreatesACrypticDonor() throws Exception {
         final SplicingVariantAlleleEvaluation ve = VariantsForTesting.HBBcodingExon1UpstreamCryptic(rd, annotator);
-        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction());
+        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction().getFeatureMap());
         assertThat(ctx, is(VisualizationContext.CRYPTIC_DONOR));
     }
 
@@ -69,28 +72,28 @@ public class VisualizationContextSelectorTest {
     @Test
     public void acceptorQuidVariant() throws Exception {
         final SplicingVariantAlleleEvaluation ve = VariantsForTesting.VWFAcceptorExon26minus2QUID(rd, annotator);
-        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction());
+        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction().getFeatureMap());
         assertThat(ctx, is(VisualizationContext.CANONICAL_ACCEPTOR));
     }
 
     @Test
     public void acceptorNonQuidVariantThatDisruptsTheSite() throws Exception {
         final SplicingVariantAlleleEvaluation ve = VariantsForTesting.TSC2AcceptorExon11Minus3(rd, annotator);
-        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction());
+        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction().getFeatureMap());
         assertThat(ctx, is(VisualizationContext.CANONICAL_ACCEPTOR));
     }
 
     @Test
     public void acceptorNonQuidVariantThatCreatesACrypticAcceptor() throws Exception {
         final SplicingVariantAlleleEvaluation ve = VariantsForTesting.COL4A5AcceptorExon11Minus8(rd, annotator);
-        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction());
+        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction().getFeatureMap());
         assertThat(ctx, is(VisualizationContext.CRYPTIC_ACCEPTOR));
     }
 
     @Test
     public void codingNonQuidVariantThatCreatesACrypticAcceptor() throws Exception {
         final SplicingVariantAlleleEvaluation ve = VariantsForTesting.RYR1codingExon102crypticAcceptor(rd, annotator);
-        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction());
+        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction().getFeatureMap());
         assertThat(ctx, is(VisualizationContext.CRYPTIC_ACCEPTOR));
     }
 
@@ -99,7 +102,7 @@ public class VisualizationContextSelectorTest {
     @Test
     public void sreVariant() throws Exception {
         final SplicingVariantAlleleEvaluation ve = VariantsForTesting.NF1codingExon9coding_SRE(rd, annotator);
-        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction());
+        final VisualizationContext ctx = selector.selectContext(ve.getPrimaryPrediction().getFeatureMap());
         assertThat(ctx, is(VisualizationContext.SRE));
     }
 }